removed basecr config due to spring errors
